+++ src/Handyman.DataContractValidator/src/TypeInfoResolvers/DictionaryTypeInfoResolver.cs
@@ -9,7 +9,7 @@ namespace Handyman.DataContractValidator.TypeInfoResolvers
 {
     internal class DictionaryTypeInfoResolver : ITypeInfoResolver
     {
-        public bool TryResolveTypeInfo(object o, TypeInfoResolverContext context, out TypeInfo typeInfo)
+        public ITypeInfo ResolveTypeInfo(object o, TypeInfoResolverContext context, Func<object, ITypeInfo> next)
         {
             var isType = o is Type;
             var type = o as Type ?? o.GetType();
@@ -18,30 +18,25 @@ namespace Handyman.DataContractValidator.TypeInfoResolvers
             {
                 var genericArguments = @interface.GetGenericArguments();
 
-                typeInfo = new DictionaryTypeInfo
+                return new DictionaryTypeInfo
                 {
                     Key = context.GetTypeInfo(genericArguments[0]),
                     Value = context.GetTypeInfo(genericArguments[1])
                 };
-
-                return true;
             }
 
             if (!isType && o is Hashtable hashtable)
             {
                 var entry = hashtable.OfType<DictionaryEntry>().Single();
 
-                typeInfo = new DictionaryTypeInfo
+                return new DictionaryTypeInfo
                 {
                     Key = context.GetTypeInfo(entry.Key),
                     Value = context.GetTypeInfo(entry.Value)
                 };
-
-                return true;
             }
 
-            typeInfo = null;
-            return false;
+            return next(o);
         }
     }
 }