+++ src/Handyman.DataContractValidator/src/TypeInfoResolvers/ValueTypeInfoResolver.cs
@@ -6,31 +6,28 @@ namespace Handyman.DataContractValidator.TypeInfoResolvers
 {
     internal class ValueTypeInfoResolver : ITypeInfoResolver
     {
-        public bool TryResolveTypeInfo(object o, TypeInfoResolverContext context, out TypeInfo typeInfo)
+        public ITypeInfo ResolveTypeInfo(object o, TypeInfoResolverContext context, Func<object, ITypeInfo> next)
         {
             if (o is Type type)
             {
                 if (type == typeof(string))
                 {
-                    typeInfo = new ValueTypeInfo { Type = typeof(string) };
-                    return true;
+                    return new ValueTypeInfo { Type = typeof(string) };
                 }
 
                 if (type.IsValueType)
                 {
                     var isNullable = type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
 
-                    typeInfo = new ValueTypeInfo
+                    return new ValueTypeInfo
                     {
                         IsNullable = isNullable,
                         Type = isNullable ? type.GetGenericArguments().Single() : type
                     };
-                    return true;
                 }
             }
 
-            typeInfo = null;
-            return false;
+            return next(o);
         }
     }
 }